Refactor conflict field validator
1. Rename 'ConflictField' annotation to 'ConflictsWith'.
2. Change annotation variable from 'source' and 'conflict' to value, which just accepts a list of conflicting field names.
3. Remove 'ConflictFields' annotation as the new 'ConflictsWith' annotation is able to accept multiple field names on its own, so no need for handling multiple annotations
4. Refactor the validation processor 'ConflictsWithValidator' to use the diffable parameter.
