--- core/src/main/java/gyro/core/validation/ConflictFields.java
@@ -1,15 +1,0 @@
-package gyro.core.validation;
-
-import java.lang.annotation.Documented;
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@Documented
-@Retention(RetentionPolicy.RUNTIME)
-@Target(ElementType.METHOD)
-@ValidatorClass(ConflictFieldsValidator.class)
-public @interface ConflictFields {
-    ConflictField[] value();
-}
